% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getConjugationsEnglish.R
\name{getConjugationsEnglish}
\alias{getConjugationsEnglish}
\title{Get Conjugations}
\usage{
getConjugationsEnglish(infinitive)
}
\arguments{
\item{infinitive}{A character string giving the verb for which all the
conjugations are desired.}
}
\value{
A data.frame containing three columns: the tense, the person,
and the conjugation of the verb.
}
\description{
Get Conjugations
}

