% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{interface}
\alias{interface}
\title{Interface}
\usage{
interface()
}
\value{
Nothing, but files are modified and the quizzes ran.
}
\description{
This is the main function that runs the program.
}

