% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/italian2English.R
\name{italian2English}
\alias{italian2English}
\title{Italian to English}
\usage{
italian2English(person, tense, verb)
}
\value{
The english verb
}
\description{
This function converts an Italian verb to an English verb.
}

