% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/startVerb.R
\name{startVerb}
\alias{startVerb}
\title{Start Verb Quiz}
\usage{
startVerb()
}
\value{
The user name, as a character string.
}
\description{
This function gets the user name which then defines the file needed to run.
}

