% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapVerbQuiz.R
\name{wrapVerbQuiz}
\alias{wrapVerbQuiz}
\title{Single Verb Quiz}
\usage{
wrapVerbQuiz(name)
}
\arguments{
\item{name}{The username provided previously.  This is used to specify what
the verbFile should be saved as (i.e. the file name).}
}
\description{
This function uses the globally defined verbFile object to run the verb
conjugation quiz.  The verbFile should be a data.table with 5 columns:
category (either person, tense, or verb), word, active, attempts, and
successes.  category is needed as the one file contains the persons (io,
tu, ...) to be tested, the verbs to be conjugated, and the tenses to use.
}

